+++ src/mdacli/cli.py
@@ -12,6 +12,7 @@ import sys
 import traceback
 import warnings
 
+import argcomplete
 from MDAnalysis.analysis.base import AnalysisBase
 from threadpoolctl import threadpool_limits
 
@@ -89,6 +90,9 @@ def cli(
 
     ap = init_base_argparse(name=name, version=version, description=description)
 
+    setup_clients(ap, title=f"{name} Analysis Modules", members=modules)
+    argcomplete.autocomplete(ap)
+
     if len(sys.argv) < 2:
         ap.error("A subcommand is required.")
 
@@ -97,7 +101,6 @@ def cli(
     #    i.e. for `mda RMSD` only the RMSD client should be build.
     # 2. for something like `mdacli -h` We do not have to build every
     #   sub parser in complete detail.
-    setup_clients(ap, title=f"{name} Analysis Modules", members=modules)
 
     args = ap.parse_args()